unfubar the bad casting i done did
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@9127 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/BackgroundRenderer.cxx
@@ -359,9 +359,9 @@ void			BackgroundRenderer::setCelestial(
   // make pretransformed display list for sun
   sunXFormList.begin();
     glPushMatrix();
-    glRotatef(atan2f(sunDirection[1], (float)(sunDirection[0]) * 180.0 / M_PI),
+    glRotatef((GLfloat)(atan2f(sunDirection[1], (sunDirection[0])) * 180.0 / M_PI),
 							0.0f, 0.0f, 1.0f);
-    glRotatef(asinf((float)(sunDirection[2]) * 180.0 / M_PI), 0.0f, -1.0f, 0.0f);
+    glRotatef((GLfloat)(asinf(sunDirection[2]) * 180.0 / M_PI), 0.0f, -1.0f, 0.0f);
     sunList.execute();
     glPopMatrix();
   sunXFormList.end();
@@ -374,7 +374,7 @@ void			BackgroundRenderer::setCelestial(
   coverage = (coverage < 0.0f) ? -sqrtf(-coverage) : coverage * coverage;
   float worldSize = BZDBCache::worldSize;
   const float moonRadius = 2.0f * worldSize *
-				atanf((float)(60.0 * M_PI / 180.0) / 60.0);
+				atanf((float)((60.0 * M_PI / 180.0) / 60.0));
   // limbAngle is dependent on moon position but sun is so much farther
   // away that the moon's position is negligible.  rotate sun and moon
   // so that moon is on the horizon in the +x direction, then compute
@@ -390,9 +390,9 @@ void			BackgroundRenderer::setCelestial(
   int moonSegements = (int)BZDB.eval("moonSegments");
   moonList.begin();
     glPushMatrix();
-    glRotatef(atan2f(moonDirection[1], (float)(moonDirection[0]) * 180.0 / M_PI),
+    glRotatef((GLfloat)(atan2f(moonDirection[1], moonDirection[0]) * 180.0 / M_PI),
 							0.0f, 0.0f, 1.0f);
-    glRotatef(asinf((float)(moonDirection[2]) * 180.0 / M_PI), 0.0f, -1.0f, 0.0f);
+    glRotatef((GLfloat)(asinf(moonDirection[2]) * 180.0 / M_PI), 0.0f, -1.0f, 0.0f);
     glRotatef((float)(limbAngle * 180.0 / M_PI), 1.0f, 0.0f, 0.0f);
     glBegin(GL_TRIANGLE_STRIP);
     // glTexCoord2f(0,-1);
@@ -606,7 +606,7 @@ void BackgroundRenderer::drawSky(SceneRenderer& renderer, bool mirror)
   // rotate sky so that horizon-point-toward-sun-color is actually
   // toward the sun
   glPushMatrix();
-  glRotatef((atan2f(sunDirection[1], (float)(sunDirection[0]) * 180.0 + 135.0) / M_PI),
+  glRotatef((GLfloat)((atan2f(sunDirection[1], sunDirection[0]) * 180.0 + 135.0) / M_PI),
 							0.0f, 0.0f, 1.0f);
 
   // draw sky
@@ -1018,7 +1018,7 @@ void			BackgroundRenderer::doInitDisplayLists()
   // sun first.  sun is a disk that should be about a half a degree wide
   // with a normal (60 degree) perspective.
   const float worldSize = BZDBCache::worldSize;
-  const float sunRadius = (float)(2.0 * worldSize * atanf(60.0*M_PI/180.0) / 60.0);
+  const float sunRadius = (float)(2.0 * worldSize * atanf((float)(60.0*M_PI/180.0)) / 60.0);
   sunList.begin();
     glBegin(GL_TRIANGLE_FAN);
       glVertex3f(2.0f * worldSize, 0.0f, 0.0f);
